보드 및 타일 자료구조 변경

+++ src/entity/defines.rs
@@ -1,7 +1,7 @@
 use bevy::prelude::*;
 
-pub const TILE_WIDTH: u8 = 4;
-pub const TILE_HEIGHT: u8 = 4;
+pub const TILE_WIDTH: usize = 4;
+pub const TILE_HEIGHT: usize = 4;
 
 #[derive(Debug, Clone, Copy, PartialEq)]
 pub enum TileState {
